feature: Validation added for crud albums and photos

+++ AlbumsAPI/Controllers/AlbumsCrudController.cs
@@ -2,6 +2,7 @@
 using AlbumsAPI.DTOs;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Primitives;
+using System.ComponentModel;
 
 namespace AlbumsAPI.Controllers
 {
@@ -42,6 +43,15 @@ namespace AlbumsAPI.Controllers
         {
             try
             {
+                if (request.Id == 0 || request.UserId == 0)
+                    throw new InvalidEnumArgumentException("Invalid id for an album or user id for user.");
+
+                if (await _albumsService.ValidateAlbumUserExistsAsync(request.Id, request.UserId))
+                {
+                    _logger.LogError($"Album already exist for given user.");
+                    throw new InvalidEnumArgumentException(nameof(request.UserId));
+                }
+
                 await _albumsService.CreateAlbumAsync(request.Id, request.UserId, request.Title);
                 return Ok();
             }
@@ -58,6 +68,15 @@ namespace AlbumsAPI.Controllers
         {
             try
             {
+                if (request.Id == 0 || request.UserId == 0)
+                    throw new InvalidEnumArgumentException("Invalid id for an album or user id for user.");
+
+                if (!await _albumsService.ValidateAlbumUserExistsAsync(request.Id, request.UserId))
+                {
+                    _logger.LogError($"Album does not exist for given user., not update was done.");
+                    throw new InvalidEnumArgumentException(nameof(request.UserId));
+                }
+
                 await _albumsService.UpdateAlbumAsync(request.Id, request.UserId, request.Title);
                 return Ok();
             }
@@ -74,6 +93,9 @@ namespace AlbumsAPI.Controllers
         {
             try
             {
+                if(id == 0)
+                    throw new InvalidEnumArgumentException("Invalid id for an album.");
+
                 await _albumsService.DeleteAlbumAsync(id);
                 return Ok();
             }
